fix(loader): decouple FP8 quant from rtp-kernel
